+++ packages/excalidraw/actions/types.ts
@@ -61,7 +61,6 @@ export type ActionName =
   | "objectsSnapMode"
   | "arrowBinding"
   | "midpointSnapping"
-  | "zoomWithScrollWheel"
   | "stats"
   | "changeStrokeColor"
   | "changeBackgroundColor"